+++ src/components/common/number-range/number-range.ts
@@ -1,0 +1,9 @@
+/*
+ * SPDX-FileCopyrightText: 2020 The HedgeDoc developers (see AUTHORS file)
+ *
+ * SPDX-License-Identifier: AGPL-3.0-only
+ */
+
+export const createNumberRangeArray = (length: number) : number[] => {
+  return Array.from(Array(length).keys())
+}