+++ CHANGELOG.md
@@ -53,6 +53,7 @@ SPDX-License-Identifier: CC-BY-SA-4.0
 - Code blocks have a 'Copy code to clipboard' button.
 - Code blocks with 'vega-lite' as language are rendered as [vega-lite diagrams](https://vega.github.io/vega-lite/examples/).
 - Markdown files can be imported into an existing note directly from the editor.
+- The table button in the toolbar opens an overlay where the user can choose the number of columns and rows
 
 ### Changed